Swiftlint integration (#46)
Adds Swiftlint to our Podfile and build process. From here on, running a build will expose swift lint warnings and errors. This also fixes a lot of the warnings and all of the swiftlint errors that were exposed.

Note that running swiftlint autocorrect fixes a lot of issues automatically (e.g. whitespace issues).